refactor(server): extract safety_cache.py module
`load_safety_cache` / `save_safety_cache` and the SAFETY_CACHE_FILE
constant move out of server.py into their own tiny module. Both
server.py (warmup load) and ws_session.py (per-request writes once
handle_init lands there) need them, and a separate module avoids the
circular import that would otherwise force a late import.

Refs #78.
